pyproject.toml
[tool.poetry]
name = "note-search"
version = "0.0.1"
description = ""
authors = ["Maggie Mao <guanqunm@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
notesearch= "notesearch:main"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
numpy = "^1.24.2"
pandas = "^1.5.3"
tiktoken = "^0.3.0"
tabulate = "^0.9.0"
tenacity = "^8.2.2"
python-dotenv = "^1.0.0"
cohere = "^4.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.1.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
